feat: filter 리셋 기능 추가
