+++ client/src/hooks/index.ts
@@ -1,5 +1,6 @@
 import { useDailyAnimes } from "@/hooks/useDailyAnimes";
 import { useFilter } from "@/hooks/useFilter";
+import { useFinderInfiniteScroll } from "@/hooks/useFinderInfiniteScroll";
 import { useFixScroll } from "@/hooks/useFixScroll";
 import { useFooterToggle } from "@/hooks/useFooterToggle";
 import { useHeaderWhite } from "@/hooks/useHeaderWhite";
@@ -10,6 +11,7 @@ import { useValitdate } from "@/hooks/useValitdate";
 export {
   useDailyAnimes,
   useFilter,
+  useFinderInfiniteScroll,
   useFooterToggle,
   useNewTitle,
   useFixScroll,
