+++ client/src/pages/Finder/Finder.tsx
@@ -1,15 +1,11 @@
-import { useInfiniteQuery } from "@tanstack/react-query";
-import { AxiosError } from "axios";
-import { useState, useEffect } from "react";
-
 import checkbox from "@/../public/assets/svgs/checkbox.svg";
 import checkboxreverse from "@/../public/assets/svgs/checkboxreverse.svg";
 import reset from "@/../public/assets/svgs/reset.svg";
 import uncheckbox from "@/../public/assets/svgs/uncheckbox.svg";
-import { getFinderAnimes } from "@/apis";
 import { Animes } from "@/components";
 import {
   useFilter,
+  useFinderInfiniteScroll,
   useFixScroll,
   useFooterToggle,
   useNewTitle,
@@ -17,8 +13,6 @@ import {
 } from "@/hooks";
 import * as S from "@/pages/Finder/styled";
 import { filterStatus } from "@/store/features/filterSlice";
-import { FinderQuery } from "@/types/finder";
-import { Anime } from "@/types/main";
 
 const getFilterImg = (status: string): string => {
   if (status === filterStatus.READY) {
@@ -48,70 +42,9 @@ export function Finder() {
     onSubtractTag,
     onAddNotTag,
     onSubtractNotTag,
-    tansfromStrings,
-    tansfromFetchHeaders,
   } = useFilter();
-
-  const { data, hasNextPage, fetchNextPage } = useInfiniteQuery<
-    FinderQuery | Anime[],
-    AxiosError
-  >({
-    queryKey: ["Finder", ...tansfromStrings()],
-    queryFn: async ({ pageParam }) => {
-      const { genre, xGenre, tag, xTag } = tansfromFetchHeaders();
-      const data = await getFinderAnimes(pageParam, genre, xGenre, tag, xTag);
-      return data;
-    },
-    staleTime: 60000 * 60,
-    refetchOnWindowFocus: false,
-    getNextPageParam: (lastPage, allPages) => {
-      if (!lastPage) {
-        return undefined;
-      }
-
-      if ((lastPage as FinderQuery).items) {
-        return (lastPage as FinderQuery).items[
-          (lastPage as FinderQuery).items.length - 1
-        ]._id;
-      }
-
-      if (!(lastPage as Anime[]).length) {
-        return undefined;
-      }
-
-      return (lastPage as Anime[])[(lastPage as Anime[]).length - 1]._id;
-    },
-  });
-
-  const [target, setTarget] = useState<HTMLDivElement | null>(null);
-
-  useEffect(() => {
-    let observer: IntersectionObserver;
-
-    if (!target) {
-      return;
-    }
-
-    if (target) {
-      observer = new IntersectionObserver(onIntersect);
-      observer.observe(target);
-    }
-
-    return () => {
-      observer.unobserve(target);
-    };
-  }, [target]);
-
-  const onIntersect = (
-    entries: IntersectionObserverEntry[],
-    observer: IntersectionObserver,
-  ) => {
-    entries.forEach(entry => {
-      if (entry.intersectionRatio > 0 && entry.isIntersecting) {
-        fetchNextPage();
-      }
-    });
-  };
+  const { data, setTarget, handleResetClick, transformAnimes } =
+    useFinderInfiniteScroll();
 
   const handleFilterClick = (id: number, status: string, type: string) => {
     if (status === filterStatus.READY) {
@@ -142,20 +75,7 @@ export function Finder() {
     return onUpdateReady(id, type);
   };
 
-  const transformAnimes = (data: any) => {
-    if (!data) {
-      return;
-    }
-
-    const temp = data.pages[0].items;
-
-    if (data.pages.length > 1) {
-      return [...temp, ...data.pages[1]];
-    }
-    return temp;
-  };
-
-  const mapedFilter = filters.map((el, index) => {
+  const mapedFilter = filters.map(el => {
     const mapedTag = el.item.map((item, filterIndex) => {
       return (
         <S.FilterItem
@@ -184,7 +104,7 @@ export function Finder() {
           <S.Filter>
             <S.Title>필터</S.Title>
             <S.ResetWrapper>
-              <S.Reset>
+              <S.Reset onClick={handleResetClick}>
                 <S.ResetText>초기화</S.ResetText>
                 <S.ResetIcon src={reset} alt='필터 초기화' />
               </S.Reset>
@@ -194,7 +114,9 @@ export function Finder() {
         </S.Bar>
         <S.AnimesContainer>
           <Animes animes={transformAnimes(data)} isFinder />
-          <div ref={setTarget} />
+          <div ref={setTarget} style={{ width: "100%", height: "80px" }}>
+            Bar
+          </div>
         </S.AnimesContainer>
       </S.Wrapper>
     </S.Container>
