+++ client/src/types/finder.ts
@@ -1,6 +1,12 @@
-import { Anime } from "@/types/main";
+interface FinderAnime {
+  _id: string;
+  image: string;
+  isAdult: boolean;
+  isOnly: boolean;
+  title: string;
+}
 
 export interface FinderQuery {
   count?: number;
-  items: Anime[];
+  items: FinderAnime[];
 }
